Moved SPI manager
